+++ tests/testthat/test_ts.R
@@ -14,19 +14,19 @@ x.thresh.fit <- fit.gpd( x.thresh, error.estimation = "none",
                         silent = TRUE )
 
 test_that( "aic accepts the right input and produces the right output", {
-  expect_equal( aic( x.block.fit ), 447.213063 )
-  expect_equal( aic( x.thresh.fit ), 1454.036538 )
+  expect_equal( aic( x.block.fit ), 455.866851 )
+  expect_equal( aic( x.thresh.fit ), 1486.3227 )
   expect_equal( aic( list( x.block.fit, x.thresh.fit ) ),
-               c( 447.213063, 1454.036538 ) )
+               c( 455.866851, 1486.322704 ) )
   expect_error( aic( x.block ) )
   expect_error( aic( as.numeric( x.block ) ) )
   expect_error( aic( list( x.block, as.numeric( x.block ) ) ) )
 })
 test_that( "bic accepts the right input and produces the right output", {
-  expect_equal( bic( x.block.fit ), 455.673908 )
-  expect_equal( bic( x.thresh.fit ), 1461.79201 )
+  expect_equal( bic( x.block.fit ), 464.375697 )
+  expect_equal( bic( x.thresh.fit ), 1494.11701 )
   expect_equal( bic( list( x.block.fit, x.thresh.fit ) ),
-               c( 455.673908, 1461.792010 ) )
+               c( 464.375697, 1494.117012 ) )
   expect_error( bic( x.block ) )
   expect_error( bic( as.numeric( x.block ) ) )
   expect_error( bic( list( x.block, as.numeric( x.block ) ) ) )
@@ -41,33 +41,33 @@ x.removed.year.end <-
 x.removed.multiple <- x.removed.year.end[
     -c( 3402, 10292, 12000, 32410 ) ]
 test_that( "remove.incomplete.years works as expected", {
-  expect_equal( length( temp.potsdam ), 45145 )
+  expect_equal( length( temp.potsdam ), 45977 )
   expect_equal( unique( year( temp.potsdam ) ),
-               seq( 1893, 2016, 1 ) )
-  expect_equal( length( x.removed ), 44924 )
+               seq( 1893, 2018, 1 ) )
+  expect_equal( length( x.removed ), 45655 )
   expect_equal( unique( year( x.removed ) ),
-               seq( 1893, 2015, 1 ) )
+               seq( 1893, 2017, 1 ) )
   expect_equal(
       unique( year( remove.incomplete.years(
           x.removed.year.beginning ) ) ),
-      c( 1893, seq( 1895, 2015, 1 ) ) )
+      c( 1893, seq( 1895, 2017, 1 ) ) )
   expect_equal(
       unique( year( remove.incomplete.years(
           x.removed.year.end ) ) ),
-      c( 1893, 1894, seq( 1896, 2015, 1 ) ) )
+      c( 1893, 1894, seq( 1896, 2017, 1 ) ) )
   expect_equal( length( remove.incomplete.years(
-      x.removed.multiple ) ), 43099 )
+      x.removed.multiple ) ), 43830 )
   expect_equal( length(
       remove.incomplete.years(
           remove.incomplete.years( 
-              x.removed.multiple ) ) ), 43099 )
+              x.removed.multiple ) ) ), 43830 )
   expect_equal( length(
-      remove.incomplete.years( x.removed[ -1 ] ) ), 44559 )
+      remove.incomplete.years( x.removed[ -1 ] ) ), 45290 )
   expect_equal(
       Reduce( c,
              lapply( remove.incomplete.years(
                  list( x.removed, x.removed[ -1 ] ) ), length ) ),
-      c( 44924, 44559 ) )
+      c( 45655, 45290 ) )
   expect_error( remove.incomplete.years(
       as.numeric( temp.potsdam ) ) )
   expect_error( remove.incomplete.years(
@@ -78,7 +78,7 @@ test_that( "remove.incomplete.years get's rid of NA in time series",{
   expect_equal( min( remove.incomplete.years( x.removed ) ),
                -16 )
   expect_equal( length( remove.incomplete.years( x.removed ) ),
-               44559 )
+               45290 )
 })
 
 test_that( "check.completeness is throwing errors on wrong input", {
@@ -105,7 +105,7 @@ test_that( "check.completeness is performing as expected", {
       1 ) } )
 
 test_that( "mode is working", {
-  expect_equal( mode( temp.potsdam ), 19.7559 - 2.08e-06 )
+  expect_equal( mode( temp.potsdam ), 19.8720993 )
 })
 
 test_that( "anomalies get its job done", {
@@ -116,16 +116,16 @@ test_that( "anomalies get its job done", {
   expect_error( anomalies( list( as.numeric( temp.potsdam ),
                                 temp.potsdam ) ) )
   expect_equal( max( anomalies( temp.potsdam ) ),
-               16.32177 + 4.19e-06 )
+               16.3111111 )
   expect_equal( min( anomalies( temp.potsdam ) ),
-               -19.12258 - 6.45e-07 )
+               -19.1206349 )
   expect_equal(
       max( Reduce( c, anomalies( list( temp.potsdam,
                                       temp.potsdam ) ) ) ),
-      16.32177 + 4.19e-06 )
+      16.3111111 )
   expect_equal(
       min( Reduce( c, anomalies( list( temp.potsdam,
                                       temp.potsdam ) ) ) ),
-               -19.12258 - 6.45e-07 )
+      -19.1206349 )
 } )
 ## End of test_ts.R