+++ NEWS.md
@@ -6,6 +6,7 @@
 - Removing GitHub part of the installation instruction in the
   README.md
 - Updating the `temp.potsdam` data set. It now covers two more years.
+- Updating all tests to cope with the new `temp.potsdam` data.
 
 # v2.0.0
 The functionality of the 1. climex core package, 2. the shiny-based